move chart related code from index to chart
