+++ src/pages/Calculator/Steps/ConfirmPlan/index.js
@@ -8,7 +8,6 @@ import useMediaQuery from "@mui/material/useMediaQuery";
 import { Typography, Button } from "@mui/material";
 import { useState, useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { PieChart, Pie, Cell, ResponsiveContainer } from "recharts";
 
 import { calculateAllConfirmPlan } from "../../../../shared/utils/calculate";
 import { handleDownload } from "./../../../../shared/utils/exportExcel";
@@ -34,46 +33,6 @@ const ConfirmPlan = ({ council }) => {
   const data = useSelector((state) => state.steps.value);
   const speciesSelection = data.speciesSelection;
   const NRCS = data.NRCS;
-
-  const poundsForPurchaseSum = speciesSelection.seedsSelected.reduce(
-    (sum, a) => sum + a.poundsForPurchase,
-    0
-  );
-
-  const plantsPerAcreSum = speciesSelection.seedsSelected.reduce(
-    (sum, a) => sum + parseFloat(a.plantsPerAcre),
-    0
-  );
-  const poundsOfSeedSum = speciesSelection.seedsSelected.reduce(
-    (sum, a) => sum + parseFloat(a.poundsOfSeed),
-    0
-  );
-  const seedsPerAcreSum = speciesSelection.seedsSelected.reduce(
-    (sum, a) => sum + parseFloat(a.seedsPerAcre),
-    0
-  );
-  const poundsOfSeedArray = [];
-  const plantsPerAcreArray = [];
-  const seedsPerAcreArray = [];
-
-  speciesSelection.seedsSelected.map((s, i) => {
-    plantsPerAcreArray.push({
-      name: s.label,
-      value: s.plantsPerAcre / plantsPerAcreSum,
-    });
-    seedsPerAcreArray.push({
-      name: s.label,
-      value: s.seedsPerAcre / seedsPerAcreSum,
-    });
-    poundsOfSeedArray.push({
-      name: s.label,
-      value: s.poundsOfSeed / poundsOfSeedSum,
-    });
-  });
-
-  const COLORS = ["#0088FE", "#00C49F", "#FFBB28", "#FF8042"];
-
-  const RADIAN = Math.PI / 180;
 
   //////////////////////////////////////////////////////////
   //                      Redux                           //
@@ -119,66 +78,6 @@ const ConfirmPlan = ({ council }) => {
   const handleModalOpen = (data) => {
     !modalOpen && setCurrentModal(data);
     setModalOpen(!modalOpen);
-  };
-  const renderCustomizedLabel = ({
-    cx,
-    cy,
-    midAngle,
-    innerRadius,
-    outerRadius,
-    percent,
-    index,
-  }) => {
-    const radius = innerRadius + (outerRadius - innerRadius) * 0.5;
-    const x = cx + radius * Math.cos(-midAngle * RADIAN);
-    const y = cy + radius * Math.sin(-midAngle * RADIAN);
-
-    return (
-      <text
-        x={x}
-        y={y}
-        fill="white"
-        textAnchor={x > cx ? "start" : "end"}
-        dominantBaseline="central"
-      >
-        {`${(percent * 100).toFixed(0)}%`}
-      </text>
-    );
-  };
-  const renderPieChart = (type) => {
-    let chartData;
-    if (type === "plantsPerAcre") {
-      chartData = plantsPerAcreArray;
-    }
-    if (type === "seedsPerAcre") {
-      chartData = seedsPerAcreArray;
-    }
-    if (type === "poundsOfSeed") {
-      chartData = poundsOfSeedArray;
-    }
-    return (
-      <ResponsiveContainer width="100%" height={200}>
-        <PieChart width={400} height={400}>
-          <Pie
-            data={chartData}
-            cx="50%"
-            cy="50%"
-            labelLine={false}
-            label={renderCustomizedLabel}
-            outerRadius={80}
-            fill="#8884d8"
-            dataKey="value"
-          >
-            {chartData.map((entry, index) => (
-              <Cell
-                key={`cell-${index}`}
-                fill={COLORS[index % COLORS.length]}
-              />
-            ))}
-          </Pie>
-        </PieChart>
-      </ResponsiveContainer>
-    );
   };
 
   const renderSeedsSelected = (seed) => {
@@ -255,10 +154,7 @@ const ConfirmPlan = ({ council }) => {
 
           <ConfirmPlanCharts
             council={council}
-            renderPieChart={renderPieChart}
-            poundsForPurchaseSum={poundsForPurchaseSum}
             speciesSelection={speciesSelection}
-            COLORS={COLORS}
             matchesMd={matchesMd}
           />
           <ConfirmPlanForm
